requirements.txt
appdirs==1.4.3
bcrypt==3.1.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
distlib==0.3.0
dnspython==1.16.0
favicon==0.7.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pipenv==2018.11.26
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
Werkzeug==1.0.1
